Rename Models to Assets and make some types internal
