+++ src/SimpleLevelEditor.State/States/Assets/Geometry.cs
@@ -1,3 +1,3 @@
-namespace SimpleLevelEditor.State.States.Models;
+namespace SimpleLevelEditor.State.States.Assets;
 
 public record Geometry(PositionTextureNormal[] Vertices, uint[] Indices);
