+++ src/SimpleLevelEditor.State/States/Assets/Map.cs
@@ -1,5 +1,5 @@
 using Detach.Parsers.Texture;
 
-namespace SimpleLevelEditor.State.States.Models;
+namespace SimpleLevelEditor.State.States.Assets;
 
 public record Map(string AbsolutePath, TextureData TextureData);
