+++ src/SimpleLevelEditor.State/States/Assets/Material.cs
@@ -1,3 +1,3 @@
-namespace SimpleLevelEditor.State.States.Models;
+namespace SimpleLevelEditor.State.States.Assets;
 
 public record Material(string Name, Map DiffuseMap);
