+++ src/SimpleLevelEditor.State/States/Assets/MaterialLibrary.cs
@@ -1,3 +1,3 @@
-namespace SimpleLevelEditor.State.States.Models;
+namespace SimpleLevelEditor.State.States.Assets;
 
 public record MaterialLibrary(string AbsolutePath, List<Material> Materials);
