+++ src/SimpleLevelEditor.State/States/Assets/Mesh.cs
@@ -1,5 +1,5 @@
 using System.Numerics;
 
-namespace SimpleLevelEditor.State.States.Models;
+namespace SimpleLevelEditor.State.States.Assets;
 
 public record Mesh(Geometry Geometry, string MaterialName, uint MeshVao, uint[] LineIndices, uint LineVao, Vector3 BoundingMin, Vector3 BoundingMax);
