+++ src/SimpleLevelEditor.State/States/Assets/PositionTextureNormal.cs
@@ -1,5 +1,5 @@
 using System.Numerics;
 
-namespace SimpleLevelEditor.State.States.Models;
+namespace SimpleLevelEditor.State.States.Assets;
 
 public record struct PositionTextureNormal(Vector3 Position, Vector2 Texture, Vector3 Normal);
