+++ src/SimpleLevelEditor.State/States/Level/LevelState.cs
@@ -2,11 +2,11 @@ using Microsoft.FSharp.Core;
 using Silk.NET.OpenGL;
 using SimpleLevelEditor.Formats;
 using SimpleLevelEditor.Formats.Types.Level;
+using SimpleLevelEditor.State.States.Assets;
 using SimpleLevelEditor.State.States.EntityConfig;
 using SimpleLevelEditor.State.States.EntityEditor;
 using SimpleLevelEditor.State.States.LevelEditor;
 using SimpleLevelEditor.State.States.Messages;
-using SimpleLevelEditor.State.States.Models;
 using SimpleLevelEditor.State.States.WorldObjectEditor;
 using SimpleLevelEditor.State.Utils;
 using System.Security.Cryptography;
