+++ src/SimpleLevelEditor/App.cs
@@ -5,7 +5,7 @@ using ImGuiNET;
 using Silk.NET.GLFW;
 using Silk.NET.OpenGL;
 using SimpleLevelEditor.State;
-using SimpleLevelEditor.State.States.Models;
+using SimpleLevelEditor.State.States.Assets;
 using SimpleLevelEditor.Ui;
 using SimpleLevelEditor.User;
 using System.Runtime.InteropServices;
