+++ src/SimpleLevelEditor/Logic/MainLogic.cs
@@ -6,11 +6,11 @@ using SimpleLevelEditor.Formats.Types.EntityConfig;
 using SimpleLevelEditor.Formats.Types.Level;
 using SimpleLevelEditor.Rendering;
 using SimpleLevelEditor.State;
+using SimpleLevelEditor.State.States.Assets;
 using SimpleLevelEditor.State.States.EntityConfig;
 using SimpleLevelEditor.State.States.EntityEditor;
 using SimpleLevelEditor.State.States.Level;
 using SimpleLevelEditor.State.States.LevelEditor;
-using SimpleLevelEditor.State.States.Models;
 using SimpleLevelEditor.State.States.WorldObjectEditor;
 using SimpleLevelEditor.Utils;
 using System.Diagnostics;
