+++ src/SimpleLevelEditor/Ui/ChildWindows/WorldObjectEditorWindow.cs
@@ -2,9 +2,9 @@ using Detach;
 using ImGuiNET;
 using SimpleLevelEditor.Formats.Types.Level;
 using SimpleLevelEditor.State;
+using SimpleLevelEditor.State.States.Assets;
 using SimpleLevelEditor.State.States.Level;
 using SimpleLevelEditor.State.States.LevelEditor;
-using SimpleLevelEditor.State.States.Models;
 using SimpleLevelEditor.State.States.WorldObjectEditor;
 
 namespace SimpleLevelEditor.Ui.ChildWindows;
