+++ src/SimpleLevelEditor/Ui/DebugWindow.cs
@@ -1,7 +1,7 @@
 using Detach;
 using ImGuiNET;
 using SimpleLevelEditor.State;
-using SimpleLevelEditor.State.States.Models;
+using SimpleLevelEditor.State.States.Assets;
 
 namespace SimpleLevelEditor.Ui;
