+++ src/SimpleLevelEditor/Ui/LevelAssetsWindow.cs
@@ -1,9 +1,9 @@
 using Detach;
 using ImGuiNET;
 using SimpleLevelEditor.State;
+using SimpleLevelEditor.State.States.Assets;
 using SimpleLevelEditor.State.States.EntityConfig;
 using SimpleLevelEditor.State.States.Level;
-using SimpleLevelEditor.State.States.Models;
 using SimpleLevelEditor.State.Utils;
 using System.Diagnostics;
